+++ packages/api/src/modules/notify/outPipes/BrevoMailNotifyPipe.ts
@@ -12,7 +12,12 @@ export enum TemplateEnum {
 }
 
 export class BrevoMailNotifyPipe extends BrevoNotifyPipe implements NotifyOutPipe {
-    accepts = [NotificationType.USER_CREATED, NotificationType.USER_FORGET_PASSWORD, NotificationType.TEST_EMAIL];
+    accepts = [
+        NotificationType.USER_CREATED,
+        NotificationType.USER_FORGET_PASSWORD,
+        NotificationType.TEST_EMAIL,
+        NotificationType.USER_CONFLICT,
+    ];
 
     private apiInstance: Brevo.TransactionalEmailsApi;