+++ packages/api/src/modules/providers/scdl/repositories/miscScdlGrant.repository.ts
@@ -24,7 +24,10 @@ export class MiscScdlGrantRepository extends MongoRepository<MiscScdlGrantEntity
         await this.collection.createIndex({ associationRna: 1 });
         await this.collection.createIndex(
             { producerId: 1, decisionReference: 1, conventionDate: 1, associationSiret: 1, object: 1, amount: 1 },
-            { unique: true },
+            {
+                unique: true,
+                name: "UNIQUE_ID_MISC_SCDL_GRANT",
+            },
         );
     }
 }